requirements.txt
crewai==0.11.0
python-docx==0.8.11
google-generativeai==0.3.1
